# GoWeb/practice3enviromentVar/docs/swagger.yaml
definitions:
  handler.request:
    properties:
      count:
        type: integer
      name:
        type: string
      price:
        type: number
      type:
        type: string
    required:
    - name
    type: object
  web.Response:
    properties:
      code:
        type: string
      data: {}
      error:
        type: string
      message:
        type: string
    type: object
info:
  contact: {}
paths:
  /products:
    get:
      consumes:
      - application/json
      description: get products
      parameters:
      - description: token
        in: header
        name: token
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/web.Response'
      summary: List products
      tags:
      - Products
    post:
      consumes:
      - application/json
      description: store products
      parameters:
      - description: token
        in: header
        name: token
        required: true
        type: string
      - description: Product to store
        in: body
        name: product
        required: true
        schema:
          $ref: '#/definitions/handler.request'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/web.Response'
      summary: Store products
      tags:
      - Products
  /products/{id}:
    delete:
      consumes:
      - application/json
      description: delete products
      parameters:
      - description: token
        in: header
        name: token
        required: true
        type: string
      - description: Product to delete
        in: body
        name: product
        schema:
          $ref: '#/definitions/handler.request'
      - description: id
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/web.Response'
      summary: Delete products
      tags:
      - Products
    patch:
      consumes:
      - application/json
      description: update products by name
      parameters:
      - description: token
        in: header
        name: token
        required: true
        type: string
      - description: name to update
        in: body
        name: product
        required: true
        schema:
          $ref: '#/definitions/handler.request'
      - description: id
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/web.Response'
      summary: Update by name the products
      tags:
      - Products
    put:
      consumes:
      - application/json
      description: update products
      parameters:
      - description: token
        in: header
        name: token
        required: true
        type: string
      - description: Product to update
        in: body
        name: product
        required: true
        schema:
          $ref: '#/definitions/handler.request'
      - description: id
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/web.Response'
      summary: Update products
      tags:
      - Products
swagger: "2.0"